% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{add_duration}
\alias{add_duration}
\title{Add time spend on webpage}
\usage{
add_duration(wt)
}
\arguments{
\item{wt}{webtrack data as data.table object}
}
\value{
data.table with the same columns as wt and a new column called duration
}
\description{
Derive the time spend on a website from the timestamps
}
