% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{agg_duration}
\alias{agg_duration}
\title{Aggregate duration of consecutive visits to a URL}
\usage{
agg_duration(wt, keep = FALSE, same_day = TRUE, duration_var = "duration")
}
\arguments{
\item{wt}{webtrack data object}

\item{keep}{logical. If number of aggregated visits should be kept as variable. Defaults to FALSE}

\item{same_day}{logical. Whether to count visits as consecutive only when on the same day. Defaults to TRUE}

\item{duration_var}{logical. Name of duration variable. Defaults to "duration".}
}
\value{
webtrack data.table with the same columns as wt with updated duration
}
\description{
Aggregate the duration of consecutive visits by a person to the same URL
}
\examples{
data("testdt_tracking")
wt <- as.wt_dt(testdt_tracking)
wt <- add_duration(wt, replace_by = "cutoff")
agg_duration(wt)
}
