% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bakshy}
\alias{bakshy}
\title{Bakshy Top500
Ideological alignment of 500 domains based on facebook data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 500 rows and 7 columns.
}
\usage{
bakshy
}
\description{
Bakshy Top500
Ideological alignment of 500 domains based on facebook data
}
\references{
Bakshy, Eytan, Solomon Messing, and Lada A. Adamic. "Exposure to ideologically diverse news and opinion on Facebook." Science 348.6239 (2015): 1130-1132.
}
\keyword{datasets}
