% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_visits}
\alias{classify_visits}
\title{Classify visits by matching to a list of classes}
\usage{
classify_visits(
  wt,
  classes,
  match_by = "domain",
  regex_on = NULL,
  return_rows_by = NULL,
  return_rows_val = NULL
)
}
\arguments{
\item{wt}{webtrack data object}

\item{classes}{a data.table containing classes that can be matched to visits.}

\item{match_by}{character. Whether to match list entry to the domain of a visit ("domain") or the host ("host") with an exact match;
or with a regular expression against the whole URL of a visit. If set to "domain" or "host",
both "wt" and "classes" need to have a column called thus. If set to "regex",
the "url" column of "wt" will be used, and you need to set "regex_on" to the
column in "classes" for which to do the regex matching. Defaults to "domain".}

\item{regex_on}{character. Column in "classes" on which to do the regex matching. Defaults to NULL.}

\item{return_rows_by}{character. A column in "classes" on which to subset the joined data.}

\item{return_rows_val}{character. The value of the columns specified in "return_rows_by", for which
data should be returned.}
}
\value{
webtrack data.table with the same columns as wt and any column in classes except the "match_by" column.
}
\description{
Classify visits by matching to a list of classes
}
\examples{
\dontrun{
data("testdt_tracking")
data("domain_list")
wt <- as.wt_dt(testdt_tracking)
wt_domains <- extract_domain(wt, drop_na = F)
wt_classes <- classify_visits(wt_domains, classes = domain_list, match_by = "domain")
domain_list$host <- domain_list$domain
wt_hosts <- extract_host(wt, drop_na = F)
wt_classes <- classify_visits(wt_hosts, classes = domain_list, match_by = "host")
data("domain_list")
regex_list = domain_list[type == "newsportals"]
wt_classes <- classify_visits(wt, classes = regex_list, match_by = "regex", regex_on = "domain")
}
}
