% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{deduplicate}
\alias{deduplicate}
\title{Deduplicate visits}
\usage{
deduplicate(
  wt,
  method = "flag",
  within = 1,
  duration_var = NULL,
  keep_nvisits = FALSE,
  same_day = TRUE
)
}
\arguments{
\item{wt}{webtrack data object}

\item{method}{character. "aggregate", "flag" or "drop". If set to "aggregate",
consecutive visits to the same URL are combined and their duration aggregated.
In this case, a duration column must be specified via "duration_var";
if "within" is specified, only
If set to "flag", duplicates within a certain time frame are flagged in a new
column called "duplicate". In this case, "within" argument must be specified.
If set to "drop", duplicates are dropped. Again, "within" argument must be specified.
Defaults to "aggregate".}

\item{within}{numeric. If method set to "flag or "drop",
a duplicate is defined only when the consecutive visit happens within
this time difference (in seconds). Defaults to 1.}

\item{duration_var}{character. Name of duration variable. Defaults to "duration".}

\item{keep_nvisits}{logical. If method set to "aggregate", this determines whether
number of aggregated visits should be kept as variable. Defaults to FALSE.}

\item{same_day}{logical. If method set to "aggregate", determines
whether to count visits as consecutive only when on the same day. Defaults to TRUE.}
}
\value{
webtrack data.table with the same columns as wt with updated duration
}
\description{
Deduplicate visits or aggregate the duration of duplicate visits
}
\details{
A duplicate visit is defined as a visit to the same URL as the previous visit,
within a certain time frame.
}
\examples{
data("testdt_tracking")
wt <- as.wt_dt(testdt_tracking)
wt <- add_duration(wt, cutoff = 300, replace_by = 300)
deduplicate(wt, method = "drop")
# deduplicate(wt, method = "aggregate", keep_nvisits = TRUE)
}
