% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domain_list}
\alias{domain_list}
\title{Domain list
classification of some domains (probably from some Stier et al. paper)}
\format{
An object of class \code{list} of length 7.
}
\usage{
domain_list
}
\description{
Domain list
classification of some domains (probably from some Stier et al. paper)
}
\keyword{datasets}
