% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{extract_host}
\alias{extract_host}
\title{Extract host from url}
\usage{
extract_host(wt, varname = "url", drop_na = TRUE)
}
\arguments{
\item{wt}{webtrack data object}

\item{varname}{character. name of the URL variable from which to extract the host. Defaults to "url".}

\item{drop_na}{boolean. Whether rows for which no host can be extracted should be dropped from the data. Defaults to TRUE.}
}
\value{
webtrack data.table with the same columns as wt and a new column called 'host' (or, if varname not equal to 'url', '\if{html}{\out{<varname>}}_host')
}
\description{
Extracts the host from urls. The host is defined as the part between the scheme (e.g., "https://") and the subdirectory.
}
\examples{
data("testdt_tracking")
wt <- as.wt_dt(testdt_tracking)
wt <- extract_host(wt)
}
