% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{extract_query}
\alias{extract_query}
\title{Extract the query from URL}
\usage{
extract_query(wt, varname = "url", decode = TRUE)
}
\arguments{
\item{wt}{webtrack data object}

\item{varname}{character. name of the column from which to extract the host. Defaults to \code{"url"}.}

\item{decode}{logical. Whether to decode the query (see \code{\link[utils:URLencode]{utils::URLdecode()}}), default to TRUE}
}
\value{
webtrack data.frame with the same columns as wt
and a new column called \code{'query'} (or, if varname not equal to \code{'url'}, \code{'<varname>_query'})
}
\description{
\code{extract_query()} adds the query string of a URL as a new column.
The query is defined as the part following a "?" after the path
but not including a fragment (anything after a "#").
}
\examples{
\dontrun{
data("testdt_tracking")
wt <- as.wt_dt(testdt_tracking)
# Extract query
wt <- extract_query(wt)
}
}
