% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_tracking}
\alias{fake_tracking}
\title{Fake data}
\format{
An object of class \code{data.frame} with 500 rows and 3 columns.
}
\usage{
fake_tracking
}
\description{
Small fake webtracking data for testing purpose
}
\keyword{datasets}
