% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_wt}
\alias{is_valid_wt}
\title{Check if a data.table is valid webtrack data}
\usage{
is_valid_wt(wt, processed = TRUE, verbose = TRUE)
}
\arguments{
\item{wt}{webtrack data as data.table object}

\item{processed}{logical. If TRUE, also checks if "duration" and "domain" are present. Otherwise just checks if the standard columns exist}

\item{verbose}{should details be printed or not}
}
\value{
logical if wt is valid webtrack data or not
}
\description{
Check if a data.table is valid webtrack data
}
