% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{news_types}
\alias{news_types}
\title{News Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 690 rows and 2 columns.
}
\usage{
news_types
}
\description{
Classification of domains into different news types
}
\references{
Stier, S., Mangold, F., Scharkow, M., & Breuer, J. (2022). Post Post-Broadcast Democracy? News Exposure in the Age of Online Intermediaries. American Political Science Review, 116(2), 768-774. doi:10.1017/S0003055421001222
}
\keyword{datasets}
