% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_dt.R
\name{print.wt_dt}
\alias{print.wt_dt}
\title{Print web tracking data}
\usage{
\method{print}{wt_dt}(x, ...)
}
\arguments{
\item{x}{object of class wt_dt}

\item{...}{additional parameters for print}
}
\description{
Print web tracking data
}
