% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{sum_activity}
\alias{sum_activity}
\title{Summarize activity per person}
\usage{
sum_activity(wt, timeframe = "date")
}
\arguments{
\item{wt}{webtrack data object.}

\item{timeframe}{character. Indicates for what time frame to aggregate visits. Possible values are "date", "week", "month", "year", or "wave".
If set to "wave", webtrack data object must contain a column call "wave". Defaults to "date".}
}
\value{
a data.table with columns "panelist_id" and column indicating the number of active time units.
}
\description{
Count number of active time periods (days, weeks, months, years, or waves) per person
}
\details{
sum_activity allows you to count the number of active days, weeks, months, years or waves by panelist_id.
A period counts as "active" if the person provided at least one visit for that period.
}
\examples{
data("testdt_tracking")
wt <- as.wt_dt(testdt_tracking)
wt_sum <- sum_activity(wt, timeframe = "date")
}
