% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{sum_durations}
\alias{sum_durations}
\title{Summarize visit duration by person}
\usage{
sum_durations(wt, var_duration = NULL, timeframe = "all", visit_class = NULL)
}
\arguments{
\item{wt}{webtrack data object.}

\item{var_duration}{character. Name of the duration variable if already present. Defaults to NULL,
in which case duration will be approximated with add_duration(wt, cutoff = 300, replace_by = "na", replace_val = NULL)}

\item{timeframe}{character. Indicates for what time frame to aggregate visits. Possible values are "date", "week", "month", "year", "wave" or "all".
If set to "wave", webtrack data object must contain a column call "wave". Defaults to "all".}

\item{visit_class}{character vector. Column(s) that contains a classification of visits.
Visits will be grouped by values in this column before being summarized. Defaults to NULL.}
}
\value{
a data.table with columns "panelist_id", column indicating the time unit unless "all" was specified,
name indicating the class variable if specified, and "duration_visits" indicating the duration of visits
(in seconds, or whatever the unit of the variable specified by "var_duration" parameter)
}
\description{
Summarize duration of visits by person within a time frame, and optionally by class of visit
}
\details{
sum_durations allows you to sum the duration of visits by panelist_id for different time periods (for example, by day).
It further allows to break down the number by any set of "classes" of visits, e.g. the type of the visit's domain.
}
\examples{
data("testdt_tracking")
wt <- as.wt_dt(testdt_tracking)
# example of visit classification
wt <- extract_domain(wt)
wt[,google:=ifelse(domain == "google.com", 1, 0)]
wt[,search:=ifelse(grepl("search", url), 1, 0)]
summary <- sum_durations(wt, var_duration = NULL, timeframe = "month", visit_class = c("google", "search"))
}
