% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_dt.R
\name{summary.wt_dt}
\alias{summary.wt_dt}
\title{Summary function for webtrack data}
\usage{
\method{summary}{wt_dt}(object, ...)
}
\arguments{
\item{object}{object of class wt_dt}

\item{...}{additional parameters for summary}
}
\value{
No return value, called for side effects
}
\description{
Summary function for webtrack data
}
