% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 45290 rows and 3 columns.
}
\usage{
test_data
}
\description{
Randomly generated webtrack data only used for illustrative purposes
}
\keyword{datasets}
