% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_survey}
\alias{test_survey}
\title{Test survey}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 9 rows and 2 columns.
}
\usage{
test_survey
}
\description{
Randomly generated survey data only used for illustrative purposes
}
\keyword{datasets}
