% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdt_survey_l}
\alias{testdt_survey_l}
\title{Test survey}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1500 rows and 7 columns.
}
\usage{
testdt_survey_l
}
\description{
Same randomly generated survey data, one row per person/wave (long format)
}
\keyword{datasets}
