% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdt_survey_w}
\alias{testdt_survey_w}
\title{Test survey}
\format{
An object of class \code{data.frame} with 500 rows and 8 columns.
}
\usage{
testdt_survey_w
}
\description{
Randomly generated survey data only used for illustrative purposes (wide format)
}
\keyword{datasets}
