% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdt_tracking}
\alias{testdt_tracking}
\title{Test data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 49612 rows and 5 columns.
}
\usage{
testdt_tracking
}
\description{
Sample of fully anomymized webtrack data from a research project with US participants
}
\keyword{datasets}
