% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vars_exist}
\alias{vars_exist}
\title{Check if columns are present}
\usage{
vars_exist(wt, vars = c("panelist_id", "url", "timestamp"))
}
\arguments{
\item{wt}{webtrack data object.}

\item{vars}{character vector of variables.
Defaults to \code{c("panelist_id", "url", "timestamp")}.}
}
\value{
A data.table object.
}
\description{
\code{vars_exist()} checks if columns are present in a webtrack data object.
By default, checks whether the data has a \code{panelist_id}, a \code{ulr} and a
\code{timestamp} column.#'
}
