% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webtrackR-package.R
\docType{package}
\name{webtrackR-package}
\alias{webtrackR}
\alias{webtrackR-package}
\title{webtrackR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/schochastics/webtrackR}
  \item Report bugs at \url{https://github.com/schochastics/webtrackR/issues}
}

}
\author{
\strong{Maintainer}: David Schoch \email{david@schochastics.net} (\href{https://orcid.org/0000-0003-2952-4812}{ORCID})

Authors:
\itemize{
  \item Frank Mangold \email{frank.mangold@gesis.org} (\href{https://orcid.org/0000-0002-9776-3113}{ORCID})
  \item Sebastian Stier \email{sebastian.stier@gesis.org} (\href{https://orcid.org/0000-0002-1217-5778}{ORCID})
}

}
\keyword{internal}
