% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{words_en}
\alias{words_en}
\title{English dictionary}
\format{
An object of class \code{data.frame} with 175393 rows and 2 columns.
}
\usage{
words_en
}
\description{
A dictionary of English-language words based on the 'Word Game Dictionary'
}
\references{
Neilson, Conor. words: List of English Words from the Scrabble Dictionary. https://cran.r-project.org/web/packages/words/index.html
}
\keyword{datasets}
