% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{add_previous_visit}
\alias{add_previous_visit}
\title{Add the previous visit as a variable}
\usage{
add_previous_visit(wt, level = "url")
}
\arguments{
\item{wt}{webtrack data object}

\item{level}{character. Either "url", "host" or "domain". Defaults to "url".}
}
\value{
webtrack data.table with the same columns as wt and a new column called url_previous, host_previous or domain_previous.
}
\description{
Adds the previous visit as a variable, either as the full url, the extracted host or the extracted domain
}
\examples{
data("testdt_tracking")
wt <- as.wt_dt(testdt_tracking)
wt <- add_previous_visit(wt, level = "url")
wt <- add_previous_visit(wt, level = "host")
wt <- add_previous_visit(wt, level = "domain")
}
