% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{add_referral}
\alias{add_referral}
\title{Identify social media referrals}
\usage{
add_referral(wt, platform_domains, patterns)
}
\arguments{
\item{wt}{webtrack data object}

\item{platform_domains}{character. A vector of platform domains for which referrers should be identified. Order and length must correspondent to 'patterns' vector.}

\item{patterns}{character. A vector of patterns for which referrers should be identified. Order and length must correspondent to 'platform_domains' vector.}
}
\value{
webtrack data.table with the same columns as wt and a new column called 'referral'.
}
\description{
Identifies whether a web visit was referred to from social media. See details.
}
\details{
To identify referrals, we rely on the method described as most valid in Schmidt et al.:
When the domain preceding a visit was to the platform in question,
and the query string of the visit's URL contains a certain pattern, we count it as a referred visit.
For Facebook, the pattern has been identified by Schmidt et al. as 'fbclid=', although this can change in future.
}
\examples{
data("testdt_tracking")
wt <- as.wt_dt(testdt_tracking)
wt <- add_referral(wt, platform_domains = "facebook.com", patterns = "fbclid=")
wt <- add_referral(wt,
  platform_domains = c("facebook.com", "twitter.com"),
  patterns = c("fbclid=", "utm_source=twitter")
)
}
