% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{add_session}
\alias{add_session}
\title{Create a session variable}
\usage{
add_session(wt, cutoff)
}
\arguments{
\item{wt}{webtrack data object}

\item{cutoff}{numeric. If the consecutive visit happens later than this value (in seconds), a new browsing session is defined}
}
\value{
webtrack data.table with the same columns as wt and a new column called duration
}
\description{
Define sessions of browsing depending on aq time cutoff for inactivity
}
\examples{
data("testdt_tracking")
wt <- as.wt_dt(testdt_tracking)
wt <- add_session(wt,cutoff = 120)
}
