% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{add_title}
\alias{add_title}
\title{Download and add the "title" of a URL}
\usage{
add_title(wt)
}
\arguments{
\item{wt}{webtrack data object}
}
\value{
webtrack data.table with the same columns as wt and a new column called "title",
which will be NA if the title cannot be retrieved.
}
\description{
Gets the title of a website and adds it as a new variable. See details.
}
\details{
The title of a website (the text within the \if{html}{\out{<title>}} tag of a web sites \if{html}{\out{<head>}})
is the text that is shown onn the "tab" when looking at the website in a browser.
It can contain useful information about a URL's content and can be used, for example,
for classification purposes. Note that it may take a while to run this function for a large number of URLs.
}
\examples{
data("testdt_tracking")
wt <- as.wt_dt(testdt_tracking)
wt <- add_title(wt)
}
