% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{aggregate_duration}
\alias{aggregate_duration}
\title{Aggregate duration of consecutive visits to a website}
\usage{
aggregate_duration(wt, keep = FALSE)
}
\arguments{
\item{wt}{data.table object containing tracking data}

\item{keep}{logical. if intermediary columns should be kept or not. defaults to FALSE}
}
\value{
data.table with the same columns as wt with updated duration
}
\description{
Aggregate duration of consecutive visits to a website
}
