% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audience_networks.R
\name{audience_incidence}
\alias{audience_incidence}
\title{Create incidence matrix for two-mode networks including audiences}
\usage{
audience_incidence(wt, mode2 = "domain", cutoff = 3)
}
\arguments{
\item{wt}{webtrack data object}

\item{mode2}{character. name of column that includes the second mode (e.g.
'domain' or 'host')}

\item{cutoff}{visits below this cutoff will not be considered as a visit}
}
\value{
incidence matrix of a two-mode network
}
\description{
Create incidence matrix for two-mode networks including audiences
}
\details{
The incidence matrix is a matrix A with entries  \code{A[i,j]=1} if panelist i visited outlet j at least once.
}
\examples{
data("testdt_tracking")
wt <- as.wt_dt(testdt_tracking)
wt <- add_duration(wt)
wt <- extract_domain(wt)
audience_incidence(wt)
}
\seealso{
to create audience networks see \link{audience_network}
}
