% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{deduplicate}
\alias{deduplicate}
\title{Deduplicate visits}
\usage{
deduplicate(wt, within = 1, drop = FALSE)
}
\arguments{
\item{wt}{webtrack data object}

\item{within}{numeric. If the consecutive visit happens within than this time difference (in seconds), it is flagged as a duplicate. Defaults to 1.}

\item{drop}{numeric. If duplicate visits should be dropped (TRUE), or just flagged with a new variable (FALSE). Defaults to FALSE.}
}
\value{
webtrack data.table with the same columns as wt and, if drop set to FALSE, a new column called "duplicate"
}
\description{
Drop consecutive visits that are to the same URL within a user-defined timeframe
}
\examples{
\dontrun{
data("testdt_tracking")
wt1 <- as.wt_dt(testdt_tracking)[1:1000] # revisit with new example data
wt2 <- as.wt_dt(testdt_tracking)[1:1000]
wt <- data.table::rbindlist(list(wt1, wt2))
wt <- as.wt_dt(wt)
wt <- deduplicate(wt)
}
}
