% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domain_list}
\alias{domain_list}
\title{Domain list
classification of some domains (probably from some Stier et al. paper)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 663 rows and 2 columns.
}
\usage{
domain_list
}
\description{
Domain list
classification of some domains (probably from some Stier et al. paper)
}
\keyword{datasets}
