% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domain_list}
\alias{domain_list}
\title{Domain list
classification of domains into news,portals, search, and social media}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 663 rows and 2 columns.
}
\usage{
domain_list
}
\description{
Domain list
classification of domains into news,portals, search, and social media
}
\references{
Stier, S., Mangold, F., Scharkow, M., & Breuer, J. (2022). Post Post-Broadcast Democracy? News Exposure in the Age of Online Intermediaries. American Political Science Review, 116(2), 768-774.
}
\keyword{datasets}
