% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{drop_query}
\alias{drop_query}
\title{Drop query/fragment from URL}
\usage{
drop_query(wt)
}
\arguments{
\item{wt}{webtrack data object}
}
\value{
webtrack data.table with the same columns as wt and a new column called url_noquery
}
\description{
Drops the query and fragment from a URL, i.e., everything after a "?" or "#"
}
\examples{
data("testdt_tracking")
wt <- as.wt_dt(testdt_tracking)
wt <- drop_query(wt)
}
