% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{extract_domain}
\alias{extract_domain}
\title{Extract domain from url}
\usage{
extract_domain(wt)
}
\arguments{
\item{wt}{webtrack data object}
}
\value{
webtrack data.table with the same columns as wt and a new column called domain
}
\description{
Extracts the domain from urls. We define the domain (e.g., "google.com") as the sum of the suffix (e.g., ".com") and the part before that and the preceding dot (e.g., "google" in "https://mail.google.com).
}
\examples{
data("test_data")
wt <- as.wt_dt(test_data)
wt <- extract_domain(wt)
}
