% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{extract_domain}
\alias{extract_domain}
\title{Extract domain from url}
\usage{
extract_domain(wt)
}
\arguments{
\item{wt}{webtrack data object}
}
\value{
webtrack data.table with the same columns as wt and a new column called domain
}
\description{
Extracts the domain and subdomain from the urls
}
