% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{extract_host}
\alias{extract_host}
\title{Extract host from url}
\usage{
extract_host(wt)
}
\arguments{
\item{wt}{webtrack data object}
}
\value{
webtrack data.table with the same columns as wt and a new column called domain
}
\description{
Extracts the host from urls. The host is defined as the part between the scheme (e.g., "https://") and the subdirectory.
}
\examples{
data("testdt_tracking")
wt <- as.wt_dt(testdt_tracking)
wt <- extract_host(wt)
}
