% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{extract_path}
\alias{extract_path}
\title{Extract path from url}
\usage{
extract_path(wt, varname = "url")
}
\arguments{
\item{wt}{webtrack data object}

\item{varname}{name of the URL variable from which to extract the domain Defaults to "url".}
}
\value{
webtrack data.table with the same columns as wt and a new column called 'path' (or, if varname not equal to 'url', '\if{html}{\out{<varname>}}_path')
}
\description{
Extracts the path from urls.
}
\examples{
data("testdt_tracking")
wt <- as.wt_dt(testdt_tracking)
wt <- extract_path(wt)
}
