% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{isolation_index}
\alias{isolation_index}
\title{Isolation Index
In terms of news exposure, the isolation index captures the extent to which conservatives disproportionately visit outlets whose other visitors are conservative}
\usage{
isolation_index(left, right)
}
\arguments{
\item{left}{vector (usually corresponds to a column in a webtrack data.table)}

\item{right}{vector (usually corresponds to a column in a webtrack data.table)}
}
\description{
Isolation Index
In terms of news exposure, the isolation index captures the extent to which conservatives disproportionately visit outlets whose other visitors are conservative
}
\references{
Cutler, David M., Edward L. Glaeser, and Jacob L. Vigdor. "The rise and decline of the American ghetto." Journal of political economy 107.3 (1999): 455-506.
Gentzkow, Matthew, and Jesse M. Shapiro. "Ideological segregation online and offline." The Quarterly Journal of Economics 126.4 (2011): 1799-1839.
}
