% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{news_types}
\alias{news_types}
\title{News Domains}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 690 rows and 2 columns.
}
\usage{
news_types
}
\description{
probably from some Stier et al. paper
}
\keyword{datasets}
