% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{sum_durations}
\alias{sum_durations}
\title{Summarize visit duration by person}
\usage{
sum_durations(wt, var_duration = NULL, timeframe = NULL, visit_class = NULL)
}
\arguments{
\item{wt}{webtrack data object.}

\item{var_duration}{character. Name of the duration variable if already present. Defaults to NULL,
in which case duration will be approximated with add_duration(wt, cutoff = 300, replace_by = "na", replace_val = NULL)}

\item{timeframe}{character. indicates for what time frame to aggregate visits. Possible values are "date", "week", "month", "year", "wave" or NULL.
If set to "wave", webtrack data object must contain a column call "wave". Defaults to NULL, in which case the output contains number of visits for the entire time.}

\item{visit_class}{character. Column that contains a classification of visits. For each value in this column,
the output will have a column indicating the number of visits belonging to that value. Defaults to NULL.}
}
\value{
a data.table with columns "panelist_id", column indicating the time unit (unless 'timeframe' set ot NULL),
"duration_visits" indicating the duration of visits (in seconds, or whatever the unit of the variable specified by "var_duration" parameter),
and a column for each value of the class (if 'visit_class' not set to NULL).
}
\description{
Summarize duration of visits by person within a time frame, and optionally by class of visit
}
\details{
sum_durations allows you to sum the duration of visits by panelist_id for different time periods (for example, by day).
It further allows to break down the number by any "classes" of visits.
Note:
\itemize{
\item If for a time frame all rows are NA on the duration column, the summarized duration for that time frame will be NA.
\item If only some of the rows of a time frame are NA on the duration column, the function will ignore those NA rows.
\item If there were no visits to a class (i.e., a value of the 'visit_class' column) for a time frame, the summarized duration for that time frame will be zero; if there were visits, but NA on duration, the summarized duration will be NA.
}
}
\examples{
data("testdt_tracking")
wt <- as.wt_dt(testdt_tracking)
wt <- extract_domain(wt, drop_na = TRUE)
wt[, google := ifelse(domain == "google.com", 1, 0)]
wt_summ <- sum_durations(wt)
wt_summ <- sum_durations(wt, var_duration = NULL, timeframe = "week", visit_class = NULL)
wt_summ <- sum_durations(wt, var_duration = NULL, timeframe = "week", visit_class = "google")
}
