% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{sum_visits}
\alias{sum_visits}
\title{Summarize number of visits by person}
\usage{
sum_visits(wt, timeframe = NULL, visit_class = NULL)
}
\arguments{
\item{wt}{webtrack data object.}

\item{timeframe}{character. indicates for what time frame to aggregate visits. Possible values are "date", "week", "month", "year", "wave" or NULL.
If set to "wave", webtrack data object must contain a column call "wave". Defaults to NULL, in which case the output contains number of visits for the entire time.}

\item{visit_class}{character. Column that contains a classification of visits. For each value in this column,
the output will have a column indicating the number of visits belonging to that value. Defaults to NULL.}
}
\value{
a data.table with columns "panelist_id", column indicating the time unit (unless 'timeframe' set ot NULL),
"n_visits" indicating the number of visits, and a column for each value of the class (if 'visit_class' not set to NULL).
}
\description{
Summarize number of visits by person within a time frame, and optionally by class of visit
}
\details{
sum_visits allows you to summarize the number of visits by panelist_id for different time periods (for example, by day).
It further allows to break down the number by any "classes" of visits.
}
\examples{
\dontrun{
data("testdt_tracking")
wt <- as.wt_dt(testdt_tracking)
wt <- extract_domain(wt, drop_na = T)
wt[, google := ifelse(domain == "google.com", 1, 0)]
wt_summ <- sum_visits(wt, timeframe = NULL, visit_class = NULL)
wt_summ <- sum_visits(wt, timeframe = "week", visit_class = NULL)
wt_summ <- sum_visits(wt, timeframe = "week", visit_class = "google")
}
}
