% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test Data (needs to be removed)}
\format{
An object of class \code{wt_dt} (inherits from \code{data.table}, \code{data.frame}) with 100 rows and 7 columns.
}
\usage{
test_data
}
\description{
Test Data (needs to be removed)
}
\keyword{datasets}
