% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vars_exist}
\alias{vars_exist}
\title{Check if columns are present}
\usage{
vars_exist(wt, vars = c("panelist_id", "url", "timestamp"))
}
\arguments{
\item{wt}{webtrack data as data.table object}

\item{vars}{character vector of variables}
}
\value{
A data.table object
}
\description{
checks if the required columns are present in the webtrack data
}
