% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webtrackR-package.R
\docType{package}
\name{webtrackR-package}
\alias{webtrackR}
\alias{webtrackR-package}
\title{webtrackR: Preprocessing and Analyzing Web Tracking Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The package implements data structures and methods to work with web tracking data. The functions cover data preprocessing steps, enriching web tracking data with external information and methods for the analysis of digital behavior as used in several academic papers (e.g., Clemm von Hohenberg et al., 2023 \doi{10.17605/OSF.IO/M3U9P}; Stier et al., 2022 \doi{10.1017/S0003055421001222}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/schochastics/webtrackR}
  \item \url{https://schochastics.github.io/webtrackR/}
  \item Report bugs at \url{https://github.com/schochastics/webtrackR/issues}
}

}
\author{
\strong{Maintainer}: David Schoch \email{david@schochastics.net} (\href{https://orcid.org/0000-0003-2952-4812}{ORCID})

Authors:
\itemize{
  \item Bernhard Clemm von Hohenberg \email{bernhard.clemm@gesis.org} (\href{https://orcid.org/0000-0002-6976-9745}{ORCID})
  \item Frank Mangold \email{frank.mangold@gesis.org} (\href{https://orcid.org/0000-0002-9776-3113}{ORCID})
  \item Sebastian Stier \email{sebastian.stier@gesis.org} (\href{https://orcid.org/0000-0002-1217-5778}{ORCID})
}

}
\keyword{internal}
