% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webtrackR-package.R
\docType{package}
\name{webtrackR-package}
\alias{webtrackR}
\alias{webtrackR-package}
\title{webtrackR: Analysing Web Tracking Data and Online News Behaviour}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements data structures and methods to work with web tracking data. This includes data preprocessing steps, methods to construct audience networks \doi{10.1080/19312458.2020.1724274}, and metrics of news audience polarization \doi{10.1080/19312458.2022.2085249}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/schochastics/webtrackR}
  \item Report bugs at \url{https://github.com/schochastics/webtrackR/issues}
}

}
\author{
\strong{Maintainer}: David Schoch \email{david@schochastics.net} (\href{https://orcid.org/0000-0003-2952-4812}{ORCID})

Authors:
\itemize{
  \item Frank Mangold \email{frank.mangold@gesis.org} (\href{https://orcid.org/0000-0002-9776-3113}{ORCID})
  \item Sebastian Stier \email{sebastian.stier@gesis.org} (\href{https://orcid.org/0000-0002-1217-5778}{ORCID})
}

}
\keyword{internal}
