% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_dt.R
\name{wt_dt}
\alias{wt_dt}
\alias{as.wt_dt}
\alias{is.wt_dt}
\title{An S3 class, based on \link{data.table}, to store webtrack data}
\usage{
as.wt_dt(
  x,
  timestamp_format = "\%Y-\%m-\%d \%H:\%M:\%OS",
  varnames = c(panelist_id = "panelist_id", url = "url", timestamp = "timestamp")
)

is.wt_dt(x)
}
\arguments{
\item{x}{data.frame containing the needed set of variables (panelist_id,url
and timestamp) but not necessarily named as such.}

\item{timestamp_format}{string. Specify the raw timestamp's formatting. Defaults to "\%Y-\%m-\%d \%H:\%M:\%OS".}

\item{varnames}{named vector of variable names that contain information "panelist_id", "url", "timestamp".}
}
\value{
a webtrack data object

logical. TRUE if x is a webtrack data object and FALSE otherwise
}
\description{
An S3 class, based on \link{data.table}, to store webtrack data

Convert a data.frame containing webtrack data to a wt_dt object
}
\details{
A \code{wt_dt} table is a \link{data.table}.
Therefore, it can be used by any function that would work on a \link{data.frame} or a \link{data.table}.
Most of the operation such as variable creation, subsetting and joins are inherited from the \link{data.table}
\verb{[]} operator, following the convention \code{DT[i,j,by]} (see data table package for detail).
These operations are applied on the data.
}
\examples{
data("testdt_tracking")
wt <- as.wt_dt(testdt_tracking)
is.wt_dt(wt)
}
\seealso{
\itemize{
\item \link{data.table} -- on which \code{wt_dt} is based
}
}
